refactor(plugin): move fragments to gatsby-image
