+++ packages/gatsby-transformer-cloudinary/gatsby-image/fragments.js
