+++ packages/gatsby-transformer-cloudinary/gatsby-image/index.js
@@ -1,7 +1,25 @@
+const fs = require('fs-extra');
+
 const {
   getFixedImageObject,
   getFluidImageObject,
 } = require('./get-image-objects');
+
+exports.addFragments = async ({ store, getNodesByType }) => {
+  const program = store.getState().program;
+
+  // Check if there are any CloudinaryAsset nodes. If so add fragments for CloudinaryAsset.
+  // The fragment will cause an error if there are no CloudinaryAsset nodes.
+  if (getNodesByType(`CloudinaryAsset`).length == 0) {
+    return;
+  }
+
+  // We have CloudinaryAsset nodes so let’s add our fragments to .cache/fragments.
+  await fs.copy(
+    require.resolve(`./fragments.js`),
+    `${program.directory}/.cache/fragments/cloudinary-asset-fragments.js`
+  );
+};
 
 exports.createGatsbyImageResolvers = ({ createResolvers, reporter }) => {
   const resolvers = {
